Did first unit test for ConcurrentDBTableModel, testing that the setTable() operation series leads to a successful setting of tablename. It's not as simple as just setting a property!
To avoid having to connect to db and have db thread present for setTable() operation series complete and do the setting of tablename in last function of the operation series, made a FakeDBTalkerFriend class to shortcut the operation series. Used dependency injection, via setting which DBTalkerFriend class to use in the ConcurrentDBTableModel ctor, to inject FakeDBTalkerFriend instance in the case of unit testing.